pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "drs.pypeline"
version = "0.0.2.dev1"
authors = [
  { name="DavidRodriguezSoaresCUI", email="fireblaze904+pypeline@gmail.com" },
]
description = "A simple infrastructure for asynchronous task planning and execution"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
  "drs.drslib>=0.10.0",
  "ruamel.yaml"
]

[project.urls]
"Homepage" = "https://github.com/DavidRodriguezSoaresCUI/pypeline"
"Bug Tracker" = "https://github.com/DavidRodriguezSoaresCUI/pypeline/issues"

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
  "pytest-cov>=4.0.0"
]
analyse = [
  "mypy",
  "bandit",
  "pylint",
  "flake8",
  "vulture"
]
documentation = [
  "furo>=2023",
  "Sphinx>=6.1.3"
]

[tool.setuptools.package-data]
pypeline = ["py.typed"]

[tool.pytest.ini_options]
addopts = "--import-mode=importlib --cov --cov-report html --cov-report term-missing --cov-fail-under 95"

[tool.coverage.run]
source = ["src"]
